Implement CORS configuration and enhance OAuth2 login handling

+++ src/main/java/com/thaihoc/miniinsta/config/RedisConfig.java
@@ -10,15 +10,15 @@ import org.springframework.data.redis.serializer.StringRedisSerializer;
 @Configuration
 public class RedisConfig {
     @Bean
-    public RedisTemplate<String, Integer> redisTemplate(RedisConnectionFactory connectionFactory) {
-        RedisTemplate<String, Integer> template = new RedisTemplate<>();
+    public RedisTemplate<String, Long> redisTemplate(RedisConnectionFactory connectionFactory) {
+        RedisTemplate<String, Long> template = new RedisTemplate<>();
         template.setConnectionFactory(connectionFactory);
 
         // Set key serializer to String
         template.setKeySerializer(new StringRedisSerializer());
 
-        // Set value serializer to Integer
-        template.setValueSerializer(new GenericToStringSerializer<>(Integer.class));
+        // Set value serializer to Long
+        template.setValueSerializer(new GenericToStringSerializer<>(Long.class));
 
         return template;
     }
